test: throttle function, delete lib/scroll-throttle
